Fix all 27 sign-compare warnings and cppcheck issues
- Remove -Wno-sign-compare suppression from Makefile
- Fix sign-compare warnings across 5 source files:
  * OCArray.c: bounds checking and casts for OCIndex vs uint64_t
  * OCString.c: range validation and string length comparisons
  * OCData.c: range bounds checking
  * OCDictionary.c: loop variables to uint64_t for count matching
  * OCNumber.c: prefetch_distance type consistency
- Fix cppcheck warnings in OCAutoreleasePool.c:
  * Remove useless assignments to function parameters
- Add WARNING_SUPPRESSIONS.md documentation
- All fixes maintain type safety with proper bounds validation
- Code formatted with clang-format

+++ src/OCDictionary.c
@@ -199,28 +199,28 @@ OCMutableDictionaryRef OCDictionaryCreateMutableCopy(OCDictionaryRef theDictiona
     return (OCMutableDictionaryRef)OCDictionaryCreate((const void **)theDictionary->keys, (const void **)theDictionary->values, theDictionary->count);
 }
 const void *OCDictionaryGetValue(OCDictionaryRef theDictionary, OCStringRef key) {
-    for (long index = 0; index < theDictionary->count; index++) {
+    for (uint64_t index = 0; index < theDictionary->count; index++) {
         if (OCStringCompare(theDictionary->keys[index], key, 0) == kOCCompareEqualTo)
             return theDictionary->values[index];
     }
     return NULL;
 }
 bool OCDictionaryContainsKey(OCDictionaryRef theDictionary, OCStringRef key) {
-    for (long index = 0; index < theDictionary->count; index++) {
+    for (uint64_t index = 0; index < theDictionary->count; index++) {
         if (OCStringCompare(theDictionary->keys[index], key, 0) == kOCCompareEqualTo)
             return true;
     }
     return false;
 }
 int64_t OCDictionaryIndexOfKey(OCDictionaryRef theDictionary, OCStringRef key) {
-    for (int64_t index = 0; index < theDictionary->count; index++) {
+    for (uint64_t index = 0; index < theDictionary->count; index++) {
         if (OCStringCompare(theDictionary->keys[index], key, 0) == kOCCompareEqualTo)
-            return index;
+            return (int64_t)index;
     }
     return -1;
 }
 bool OCDictionaryContainsValue(OCDictionaryRef theDictionary, const void *value) {
-    for (long index = 0; index < theDictionary->count; index++) {
+    for (uint64_t index = 0; index < theDictionary->count; index++) {
         if (theDictionary->values[index] == value)
             return true;
     }
